pyproject.toml
[project]
name = "slack-qa-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3",
    "langgraph>=1.1.8,<2.0",
    "langchain-openai>=1.1,<2.0",
    "python-dotenv>=1.0",
    "slack-bolt>=1.21",
    "slackify-markdown>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[project.scripts]
slack-agent = "src.cli.slack:main"
ask = "src.cli.ask:main"
trace = "src.cli.trace:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
